Move Work type to sub folder
